+++ HTTP_WebServer/Socket.cpp
@@ -1,0 +1,48 @@
+#define _WINSOCK_DEPRECATED_NO_WARNINGS
+#include "Declarations.h"
+
+using namespace std;
+
+WSAData InitWinsock()
+{
+	WSAData wsaData;
+
+	if (NO_ERROR != WSAStartup(MAKEWORD(2, 2), &wsaData))
+	{
+		cout << "Web Server: Error at WSAStartup()\n";
+		exit(0);
+	}
+
+	return wsaData;
+}
+
+SOCKET CreateSocket()
+{
+	SOCKET listen_socket = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
+	if (INVALID_SOCKET == listen_socket)
+	{
+		cout << "Web Server: Error at socket(): " << WSAGetLastError() << endl;
+		WSACleanup();
+		exit(0);
+	}
+
+	return listen_socket;
+}
+
+sockaddr_in CreateSocketAdd(SOCKET& m_socket)
+{
+	sockaddr_in serverService;
+	serverService.sin_family = AF_INET;
+	serverService.sin_addr.s_addr = INADDR_ANY;
+	serverService.sin_port = htons(PORT);
+
+	if (SOCKET_ERROR == bind(m_socket, (SOCKADDR*)&serverService, sizeof(serverService)))
+	{
+		cout << "Web Server: Error at bind(): " << WSAGetLastError() << endl;
+		closesocket(m_socket);
+		WSACleanup();
+		exit(0);
+	}
+
+	return serverService;
+}
