--- HTTP_WebServer/socketsHandler.cpp
@@ -1,3 +1,0 @@
-#define _CRT_SECURE_NO_WARNINGS
-#define _WINSOCK_DEPRECATED_NO_WARNINGS
-#include "functions.h"
